Fix enlarge/ensmallen text hotkeys for Windows
